Refactor identity management to use ApplicationUser
Updated BankingSystemContext to inherit from IdentityDbContext<ApplicationUser>.
Replaced User class with ApplicationUser, adding properties like SSN, FirstName,
LastName, Address, Phone, JoinDate, BirthDate, and IsDeleted.
Changed CustomerId and ManagerId types from int to string across models.
Adjusted decimal precision in configurations and updated UserConfiguration.
Generated migration files to reflect schema changes and updated Program.cs for
identity configuration. Created ApplicationUser class for enhanced user management.

+++ BankingSystem.DAL/Models/Loan.cs
@@ -25,9 +25,9 @@ namespace BankingSystem.DAL.Models
 
     public class Loan : BaseEntity
     {
-        public double LoanAmount { get; set; }
-        public double CurrentDebt { get; set; }
-        public double InterestRate { get; set; }
+        public decimal LoanAmount { get; set; }
+        public decimal CurrentDebt { get; set; }
+        public int InterestRate { get; set; }
         public bool IsDeleted { get; set; }
         public int DurationInMonth { get; set; }
         public LoanStatus LoanStatus { get; set; }
